requirements.txt
python-dotenv==0.18.0
Django==3.2.6
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.1.0
celery==5.1.2
flower==1.0.0
redis==3.5.3
